Move sqltestutil to internal
