+++ instrumentation/database/sql/splunksql/internal/sqltestutil/go.mod
@@ -24,7 +24,7 @@ require (
 )
 
 replace (
-	github.com/signalfx/splunk-otel-go => ../../../../../
-	github.com/signalfx/splunk-otel-go/instrumentation/database/sql/splunksql => ../
-	github.com/signalfx/splunk-otel-go/instrumentation/internal => ../../../../internal/
+	github.com/signalfx/splunk-otel-go => ../../../../../../
+	github.com/signalfx/splunk-otel-go/instrumentation/database/sql/splunksql => ../../
+	github.com/signalfx/splunk-otel-go/instrumentation/internal => ../../../../../internal/
 )
